Add build of the meta-google layer
For virtio research we need DomD to host the agl_services, which are
provided by Google together with its trout device.

This patch adds usage of the 'meta-google' layer to DomD in case if we
build DomA.

Additionally, it adds additional dependency from DomD to DomA, as the
'meta-google' Yocto layer is used directly from the AOSP source tree. So
it should be fetched before DomD build is started.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>