Remove rust from build of DomD & DomU
After switching to the Yocto Kirkstone release we've identified several
minor dependencies, which include the 'rust' package into the DomD and
DomU build. That causes a significant increase in the build-time. Due
to the insignificance of the identified dependencies, we've decided to
cut them out.

Those dependencies are:
- adwaita-icon-theme, used by the 'gtk+3' package
- librsvg, used by the 'gstreamer1.0-plugins-bad' package

Operability of the fix was confirmed using the bitbake dependency
graph.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>